services/requirements.txt
# Core AI and LangChain dependencies
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-google-genai>=0.0.8
langchain-community>=0.0.12
langchain-core>=0.1.0

# TaskforgeHQ/ERPNext client
frappeclient

# Web framework and API server
flask>=2.3.0
flask-cors>=4.0.0

# Database and storage
pymongo>=4.0.0

# HTTP requests and utilities
requests>=2.28.0

# Data handling and validation
python-dotenv>=1.0.0

# Optional: Production server
gunicorn>=20.1.0

# Optional: Development tools
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0

# System and logging
python-dateutil>=2.8.0
